// src/app/shared/models/producto.interface.ts
export interface ProductoI {
    id?: string;
    codigo: string;
    nombre: string;
    unitid: string;
    precio_venta: string;
    costo_financiero: string;
    id_linea: string;
    id_categoria: string;
    codigo_id_proveedor: string;
    estado: string;
    tipo_fabricante: string;
    numero_parte: string;
    linea: string;
    categoria: string;
    uen: string;
    marca: string;
    precio_minimo: string;
    descripcion: string;
    imagen_150: string;
    imagen_450: string;
  }